# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Sistema de Matrículas",
    "version": "1.0.0",
    "description": "API para SPA de matrículas - documentação gerada via swagger-jsdoc"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "ra": {
            "type": [
              "string",
              "null"
            ]
          },
          "siape": {
            "type": [
              "string",
              "null"
            ]
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Turma": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "codigo": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "vagas": {
            "type": "integer"
          },
          "disciplina_id": {
            "type": [
              "integer",
              "null"
            ]
          },
          "professor_id": {
            "type": [
              "integer",
              "null"
            ]
          },
          "horario_id": {
            "type": [
              "integer",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "id": 1,
          "codigo": "TURMA2025A",
          "nome": "Turma de Exemplo",
          "vagas": 40,
          "disciplina_id": 2,
          "professor_id": 3,
          "horario_id": 1,
          "created_at": "2025-11-02T12:00:00.000Z"
        }
      },
...
      "UsuarioCreate": {
        "type": "object",
        "required": [
          "nome",
          "email",
          "senha"
        ],
        "properties": {
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "senha": {
            "type": "string"
          },
          "ra": {
            "type": [
              "string",
              "null"
            ]
          },
          "siape": {
            "type": [
              "string",
              "null"
            ]
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "RoleAssign": {
        "type": "object",
        "required": [
          "role"
        ],
        "properties": {
          "role": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/usuarios/register": {
      "post": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Registrar novo usuário (self-registration)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado"
          }
        }
      }
    },
    "/api/usuarios": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Listar usuários (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuários"
          }
        }
      }
    },
    "/api/usuarios/professores": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Listar usuários com role PROFESSOR (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de professores",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Usuario"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/usuarios/professores/turmas": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Lista professores com suas turmas (requer autenticação)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de professores com turmas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "nome": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "turmas": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/Turma"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/usuarios/{id}/roles": {
      "post": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Atribuir role a um usuário (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleAssign"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role atribuída"
          }
        }
      }
    },
    "/api/usuarios/{id}/turmas": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Lista as turmas atribuídas a um professor (ADMIN ou o próprio professor)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do professor"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de turmas do professor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "professor_id": {
                      "type": "integer"
                    },
                    "turmas": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Turma"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Permissão insuficiente"
          }
        }
      }
    }
  },
...
}